+++ src/constants/newsBreadcrumb.ts
@@ -1,0 +1,10 @@
+export const newsBreadcrumb = [
+  {
+    name: 'Home',
+    path: '/unou',
+  },
+  {
+    name: 'News Archive',
+    path: '/unou/news',
+  },
+];
