+++ src/constants/portfolioBreadcrumb.ts
@@ -1,0 +1,10 @@
+export const portfolioBreadcrumb = [
+  {
+    name: 'Home',
+    path: '/unou',
+  },
+  {
+    name: 'Portfolio',
+    path: '/unou/portfolio',
+  },
+];
