+++ src/constants/sanouPortfolioBreadcrumb.ts
@@ -1,0 +1,10 @@
+export const sanouPortfolioBreadcrumb = [
+  {
+    name: 'Home',
+    path: '/sanou',
+  },
+  {
+    name: 'Portfolio',
+    path: '/sanou/portfolio',
+  },
+];
