+++ src/constants/settlementBreadcrumb.ts
@@ -1,0 +1,10 @@
+export const settlementBreadcrumb = [
+  {
+    name: 'Home',
+    path: '/sanou',
+  },
+  {
+    name: '決算公告',
+    path: '/unou/news',
+  },
+];
